Add a WIP test setup for the PAM integration.
This uses pam-wrapper to enable testing a PAM stack without root and without modifying system-wide configuration. Eventually I'll also want a real integration test using NixOS' excellent support for VMs, but it's faster to develop with local testing.

This uses pam-test as the PAM application. This actually requires local patches that aren't reflected here as the upstream pam-test checks for the existence of its config under /etc/pam.d and we fake that with pam-wrapper.

This setup also hardcodes paths that I'm using for development. I'll have to figure out how to remove absolute paths elsewhere.